The attempt to give a write error to the response promise was being done in two separate listeners to the write promise. The place other than the one removed here is in ApnsClient.sendNotification.
There is also a listener added to the response promise itself inside ApnsClientHandler.write which removes the entry for the notification from the responsePromises map. So, the problem we see is that if the write promise listener in ApnsClient is executed first, it will complete the response promise, which will then trigger the removal of the response promise from the responsePromises map in ApnsClientHandler. Then the second listener on the write promise (the one we are removing) is called and when it tries to find the response promise from the map, it fails and logs errors. The listener on the write promise in ApnsClient is sufficient to propogate a write failure through to the response promise.