Drop the response promise map in favor of stream properties
Rather than having each ApnsClientHandler maintain a map of promises to
be notified when the server responds, we now attach the response
promises to individual HTTP/2 streams. This is, I think, generally
clearer as a conceptual model (responses come from the server on a
particular stream, and the relevant promise is attached to that same
stream) and provides a clearer mechanism for cleaning up when streams
close.

This also introduces a clearer "retry" mechanism for temporarily-failed
push notifications, which paves the way for a fix for #476.