Create new Auth Token before expiration of current Auth Token
Create a new Auth Tokens from the Auth key when the current Auth Token is about
to expire.  From the Apple documentation, an Auth Token is good for one hour.
If the current Auth Token is within 30 seconds of expiring or has already
expired, we create a new Auth Token.

The 30 seconds is an arbitrary fudge-factor to keep from hitting the race
condition where the token was not expired when we sent it but is expired
when the APNs server tries to process it.  It could most likely be much
less but there is no real cost to rotating the Auth Token 30 seconds early.